// gen/diff_pb.ts
export interface Symbol {
    /**
     * Name of the symbol
     *
     * @generated from protobuf field: string name = 1;
     */
    name: string;
    /**
     * Demangled name of the symbol
     *
     * @generated from protobuf field: optional string demangled_name = 2;
     */
    demangled_name?: string;
    /**
     * Symbol address
     *
     * @generated from protobuf field: uint64 address = 3;
     */
    address: bigint;
    /**
     * Symbol size
     *
     * @generated from protobuf field: uint64 size = 4;
     */
    size: bigint;
    /**
     * Bitmask of SymbolFlag
     *
     * @generated from protobuf field: uint32 flags = 5;
     */
    flags: number;
}
// ...
export interface InstructionDiff {
    /**
     * @generated from protobuf field: objdiff.diff.DiffKind diff_kind = 1;
     */
    diff_kind: DiffKind;
    /**
     * @generated from protobuf field: optional objdiff.diff.Instruction instruction = 2;
     */
    instruction?: Instruction;
    /**
     * @generated from protobuf field: optional objdiff.diff.InstructionBranchFrom branch_from = 3;
     */
    branch_from?: InstructionBranchFrom;
    /**
     * @generated from protobuf field: optional objdiff.diff.InstructionBranchTo branch_to = 4;
     */
    branch_to?: InstructionBranchTo;
    /**
     * @generated from protobuf field: repeated objdiff.diff.ArgumentDiff arg_diff = 5;
     */
    arg_diff: ArgumentDiff[];
}
/**
 * @generated from protobuf message objdiff.diff.ArgumentDiff
 */
export interface ArgumentDiff {
    /**
     * @generated from protobuf field: optional uint32 diff_index = 1;
     */
    diff_index?: number;
}
/**
 * @generated from protobuf message objdiff.diff.InstructionBranchFrom
 */
export interface InstructionBranchFrom {
    /**
     * @generated from protobuf field: repeated uint32 instruction_index = 1;
     */
    instruction_index: number[];
    /**
     * @generated from protobuf field: uint32 branch_index = 2;
     */
    branch_index: number;
}
/**
 * @generated from protobuf message objdiff.diff.InstructionBranchTo
 */
export interface InstructionBranchTo {
    /**
     * @generated from protobuf field: uint32 instruction_index = 1;
     */
    instruction_index: number;
    /**
     * @generated from protobuf field: uint32 branch_index = 2;
     */
    branch_index: number;
}
/**
 * @generated from protobuf message objdiff.diff.FunctionDiff
 */
export interface FunctionDiff {
    /**
     * @generated from protobuf field: objdiff.diff.Symbol symbol = 1;
     */
    symbol?: Symbol;
    /**
     * @generated from protobuf field: repeated objdiff.diff.InstructionDiff instructions = 2;
     */
    instructions: InstructionDiff[];
    /**
     * @generated from protobuf field: optional float match_percent = 3;
     */
    match_percent?: number;
}
// ...
export interface SectionDiff {
    /**
     * @generated from protobuf field: string name = 1;
     */
    name: string;
    /**
     * @generated from protobuf field: objdiff.diff.SectionKind kind = 2;
     */
    kind: SectionKind;
    /**
     * @generated from protobuf field: uint64 size = 3;
     */
    size: bigint;
    /**
     * @generated from protobuf field: uint64 address = 4;
     */
    address: bigint;
    /**
     * @generated from protobuf field: repeated objdiff.diff.FunctionDiff functions = 5;
     */
    functions: FunctionDiff[];
    /**
     * @generated from protobuf field: repeated objdiff.diff.DataDiff data = 6;
     */
    data: DataDiff[];
    /**
     * @generated from protobuf field: optional float match_percent = 7;
     */
    match_percent?: number;
}
/**
 * @generated from protobuf message objdiff.diff.ObjectDiff
 */
export interface ObjectDiff {
    /**
     * @generated from protobuf field: repeated objdiff.diff.SectionDiff sections = 1;
     */
    sections: SectionDiff[];
}
// ...
export enum DiffKind {
    /**
     * @generated from protobuf enum value: DIFF_NONE = 0;
     */
    DIFF_NONE = 0,
    /**
     * @generated from protobuf enum value: DIFF_REPLACE = 1;
     */
    DIFF_REPLACE = 1,
    /**
     * @generated from protobuf enum value: DIFF_DELETE = 2;
     */
    DIFF_DELETE = 2,
    /**
     * @generated from protobuf enum value: DIFF_INSERT = 3;
     */
    DIFF_INSERT = 3,
    /**
     * @generated from protobuf enum value: DIFF_OP_MISMATCH = 4;
     */
    DIFF_OP_MISMATCH = 4,
    /**
     * @generated from protobuf enum value: DIFF_ARG_MISMATCH = 5;
     */
    DIFF_ARG_MISMATCH = 5
}
/**
 * @generated from protobuf enum objdiff.diff.SectionKind
 */
export enum SectionKind {
    /**
     * @generated from protobuf enum value: SECTION_UNKNOWN = 0;
     */
    SECTION_UNKNOWN = 0,
    /**
     * @generated from protobuf enum value: SECTION_TEXT = 1;
     */
    SECTION_TEXT = 1,
    /**
     * @generated from protobuf enum value: SECTION_DATA = 2;
     */
    SECTION_DATA = 2,
    /**
     * @generated from protobuf enum value: SECTION_BSS = 3;
     */
    SECTION_BSS = 3,
    /**
     * @generated from protobuf enum value: SECTION_COMMON = 4;
     */
    SECTION_COMMON = 4
}
// @generated message type with reflection information, may provide speed optimized methods
class Symbol$Type extends MessageType<Symbol> {
    constructor() {
        super("objdiff.diff.Symbol", [
            { no: 1, name: "name", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "demangled_name", kind: "scalar", localName: "demangled_name", opt: true, T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "address", kind: "scalar", T: 4 /*ScalarType.UINT64*/, L: 0 /*LongType.BIGINT*/ },
            { no: 4, name: "size", kind: "scalar", T: 4 /*ScalarType.UINT64*/, L: 0 /*LongType.BIGINT*/ },
            { no: 5, name: "flags", kind: "scalar", T: 13 /*ScalarType.UINT32*/ }
        ]);
    }
    create(value?: PartialMessage<Symbol>): Symbol {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.name = "";
        message.address = 0n;
        message.size = 0n;
        message.flags = 0;
        if (value !== undefined)
            reflectionMergePartial<Symbol>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: Symbol): Symbol {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string name */ 1:
                    message.name = reader.string();
                    break;
                case /* optional string demangled_name */ 2:
                    message.demangled_name = reader.string();
                    break;
                case /* uint64 address */ 3:
                    message.address = reader.uint64().toBigInt();
                    break;
                case /* uint64 size */ 4:
                    message.size = reader.uint64().toBigInt();
                    break;
                case /* uint32 flags */ 5:
                    message.flags = reader.uint32();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: Symbol, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string name = 1; */
        if (message.name !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.name);
        /* optional string demangled_name = 2; */
        if (message.demangled_name !== undefined)
            writer.tag(2, WireType.LengthDelimited).string(message.demangled_name);
        /* uint64 address = 3; */
        if (message.address !== 0n)
            writer.tag(3, WireType.Varint).uint64(message.address);
        /* uint64 size = 4; */
        if (message.size !== 0n)
            writer.tag(4, WireType.Varint).uint64(message.size);
        /* uint32 flags = 5; */
        if (message.flags !== 0)
            writer.tag(5, WireType.Varint).uint32(message.flags);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message objdiff.diff.Symbol
 */
export const Symbol = new Symbol$Type();
// ...
export const Instruction = new Instruction$Type();
// ...
export const InstructionDiff = new InstructionDiff$Type();
// ...
export const ArgumentDiff = new ArgumentDiff$Type();
// ...
export const InstructionBranchFrom = new InstructionBranchFrom$Type();
// ...
export const InstructionBranchTo = new InstructionBranchTo$Type();
// @generated message type with reflection information, may provide speed optimized methods
class FunctionDiff$Type extends MessageType<FunctionDiff> {
    constructor() {
        super("objdiff.diff.FunctionDiff", [
            { no: 1, name: "symbol", kind: "message", T: () => Symbol },
            { no: 2, name: "instructions", kind: "message", repeat: 1 /*RepeatType.PACKED*/, T: () => InstructionDiff },
            { no: 3, name: "match_percent", kind: "scalar", localName: "match_percent", opt: true, T: 2 /*ScalarType.FLOAT*/ }
        ]);
    }
    create(value?: PartialMessage<FunctionDiff>): FunctionDiff {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.instructions = [];
        if (value !== undefined)
            reflectionMergePartial<FunctionDiff>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: FunctionDiff): FunctionDiff {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* objdiff.diff.Symbol symbol */ 1:
                    message.symbol = Symbol.internalBinaryRead(reader, reader.uint32(), options, message.symbol);
                    break;
                case /* repeated objdiff.diff.InstructionDiff instructions */ 2:
                    message.instructions.push(InstructionDiff.internalBinaryRead(reader, reader.uint32(), options));
                    break;
                case /* optional float match_percent */ 3:
                    message.match_percent = reader.float();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: FunctionDiff, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* objdiff.diff.Symbol symbol = 1; */
        if (message.symbol)
            Symbol.internalBinaryWrite(message.symbol, writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        /* repeated objdiff.diff.InstructionDiff instructions = 2; */
        for (let i = 0; i < message.instructions.length; i++)
            InstructionDiff.internalBinaryWrite(message.instructions[i], writer.tag(2, WireType.LengthDelimited).fork(), options).join();
        /* optional float match_percent = 3; */
        if (message.match_percent !== undefined)
            writer.tag(3, WireType.Bit32).float(message.match_percent);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message objdiff.diff.FunctionDiff
 */
export const FunctionDiff = new FunctionDiff$Type();
// ...
export const DataDiff = new DataDiff$Type();
// ...
export const SectionDiff = new SectionDiff$Type();
// ...
export const ObjectDiff = new ObjectDiff$Type();